% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSharpLayer.R
\name{layer_input}
\alias{layer_input}
\title{Construct input layer for NET# definition.}
\usage{
layer_input(shape, name = "layer_input")
}
\arguments{
\item{shape}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}}

\item{name}{Name of the layer}
}
\description{
Construct input layer for NET# definition.
}

