% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSharpLayer.R
\name{layer_norm}
\alias{layer_norm}
\title{Create response normalization layer.}
\usage{
layer_norm(layer, kernelshape, inputshape, name, inputname, stride, padding,
  mapcount, alpha, beta, activation = NA, ...)
}
\arguments{
\item{layer}{A layer object, e.g. using \code{\link{layer_input}}, or NULL}

\item{kernelshape}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}}

\item{inputshape}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}.  If \code{layer} is specified, this can be NULL.}

\item{name}{Name of the layer}

\item{inputname}{Name of the preceding layer. If \code{layer} is specified, this can be NULL.}

\item{stride}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}.}

\item{padding}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}.}

\item{mapcount}{Number of maps to create.}

\item{activation}{Activation function, e.g. \code{rlinear}}
}
\description{
Create response normalization layer.
}
\seealso{
Other layer.definition.functions: \code{\link{layer_conv}},
  \code{\link{layer_full}}, \code{\link{layer_input}},
  \code{\link{layer_output}}, \code{\link{layer_pool}}
}

