% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxlayer.R, R/netSharpLayer.R
\name{mxlayer_pool}
\alias{mxlayer_pool}
\title{Create maximum pooling layer.}
\usage{
mxlayer_pool(layer, kernelshape, inputshape, name, inputname, stride, padding,
  mapcount, activation = NA, ...)

mxlayer_pool(layer, kernelshape, inputshape, name, inputname, stride, padding,
  mapcount, activation = NA, ...)
}
\arguments{
\item{layer}{A layer object, e.g. using \code{\link{mxlayer_input}}, or NULL}

\item{kernelshape}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}}

\item{inputshape}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}.  If \code{layer} is specified, this can be NULL.}

\item{name}{Name of the layer}

\item{inputname}{Name of the preceding layer. If \code{layer} is specified, this can be NULL.}

\item{stride}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}.}

\item{padding}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}.}

\item{mapcount}{Number of maps to create.}

\item{activation}{Activation function, e.g. \code{rlinear}}
}
\description{
Create maximum pooling layer.

Create maximum pooling layer.
}
\seealso{
Other layer.definition.functions: \code{\link{mxlayer_conv}},
  \code{\link{mxlayer_full}}, \code{\link{mxlayer_input}},
  \code{\link{mxlayer_norm}}, \code{\link{mxlayer_output}}

Other layer.definition.functions: \code{\link{mxlayer_conv}},
  \code{\link{mxlayer_full}}, \code{\link{mxlayer_input}},
  \code{\link{mxlayer_norm}}, \code{\link{mxlayer_output}}
}

