% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxReadModel.R
\name{mxReadModel}
\alias{mxReadModel}
\title{Read machine learning model from file.}
\usage{
mxReadModel(file, ...)
}
\arguments{
\item{file}{a \link{connection} or the name of the file where the R object is saved to or read from.}

\item{...}{passed to \link{readRDS}}
}
\description{
Read machine learning model from file.
}

