% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxReadModel.R
\name{mxSaveModel}
\alias{mxSaveModel}
\title{Write machine learning model to file.}
\usage{
mxSaveModel(x, file = "", ...)
}
\arguments{
\item{file}{a \link{connection} or the name of the file where the R object is saved to or read from.}

\item{...}{passed to \link{saveRDS}}
}
\description{
Write machine learning model to file.
}

