% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSharpLayer.R
\name{is.netSharpLayer}
\alias{as.netSharpLayer}
\alias{is.netSharpLayer}
\title{Create or test for objects of type netSharpLayer.}
\usage{
is.netSharpLayer(x)

as.netSharpLayer(x, name, shape)
}
\description{
Create or test for objects of type netSharpLayer.
}

