% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnExtract.R
\name{nnExtractIterationImprovement}
\alias{nnExtractIterationImprovement}
\title{Extract the percentage improvement in modeling error from neuralNet captured output.}
\usage{
nnExtractIterationImprovement(model_output)
}
\arguments{
\item{model_output}{Result from \code{\link{capture.output}} on \code{\link{neuralNet}}}
}
\description{
Extract the percentage improvement in modeling error from neuralNet captured output.
}

