% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnExtract.R
\name{nnExtractIterations}
\alias{nnExtractIterations}
\title{Extract iteration data from captured output of neuralNet model.}
\usage{
nnExtractIterations(model_output)
}
\arguments{
\item{model_output}{Result from \code{\link{capture.output}} on \code{\link{neuralNet}}}
}
\description{
Extract iteration data from captured output of neuralNet model.
}

