% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnExtract.R
\name{nnExtractTotalModelWeights}
\alias{nnExtractTotalModelWeights}
\title{Extract the the total number of model weights from a neuralNet captured output.}
\usage{
nnExtractTotalModelWeights(model_output)
}
\arguments{
\item{model_output}{Result from \code{\link{capture.output}} on \code{\link{neuralNet}}}
}
\description{
Extract the the total number of model weights from a neuralNet captured output.
}

