% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnlayer.R
\name{nnlayer_pool}
\alias{nnlayer_pool}
\title{Create maximum pooling layer.}
\usage{
nnlayer_pool(layer, kernelshape, inputshape, name, inputname, stride, lowerpad,
  mapcount, activation = NA)
}
\arguments{
\item{layer}{A layer object, e.g. using \code{\link{nnlayer_input}}, or NULL}

\item{kernelshape}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}}

\item{inputshape}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}.  If \code{layer} is specified, this can be NULL.}

\item{name}{Name of the layer}

\item{inputname}{Name of the preceding layer. If \code{layer} is specified, this can be NULL.}

\item{stride}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}.}

\item{lowerpad}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}.}

\item{mapcount}{Number of maps to create.}

\item{activation}{Activation function, e.g. \code{rlinear}}
}
\description{
Create maximum pooling layer.
}
\seealso{
Other layer.definition.functions: \code{\link{nnlayer_conv}},
  \code{\link{nnlayer_full}}, \code{\link{nnlayer_input}},
  \code{\link{nnlayer_norm}}, \code{\link{nnlayer_output}}
}

