% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnReadModel.R
\name{nnReadModel}
\alias{nnDeserialize}
\alias{nnDeserializeModel}
\alias{nnReadModel}
\alias{nnSaveModel}
\alias{nnSerialize}
\alias{nnSerializeModel}
\title{Read machine learning model from file.}
\usage{
nnReadModel(file, ...)

nnSerialize(x)

nnDeserialize(x)

nnSaveModel(x, file = "", ...)

nnSerializeModel(x)

nnDeserializeModel(x)
}
\arguments{
\item{file}{a \code{\link{connection}} or the name of the file where the R object is saved to or read from.}

\item{...}{passed to \code{\link{readRDS}} and \code{\link{saveRDS}}}

\item{x}{Model object}
}
\description{
Read machine learning model from file.

Write machine learning model to file.

Serialize model to ASCII string.

Serialize model to base64 encoded string.
}

