% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructNetDefinition.R
\name{reconstructNetDefinition}
\alias{reconstructNetDefinition}
\title{Reconstruct the NET# net definition from an mxNeuralNet object}
\usage{
reconstructNetDefinition(modelObject)
}
\arguments{
\item{modelObject}{A \link{mxNeuralNet} model object}
}
\value{
A character vector containing the net definition
}
\description{
Reconstruct the NET# net definition from an mxNeuralNet object
}
\examples{
library(MicrosoftRML)
model <- mxNeuralNet(isCase ~ age + parity + education + spontaneous + induced,
                     transforms = list(isCase = case == 1),
                     data = infert,
                     numHiddenNodes = 3
)
cat(
  reconstructNetDefinition(model)
)
}

