% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnlayer_methods.R
\name{is.netSharpLayer}
\alias{as.netSharpLayer}
\alias{is.netSharpLayer}
\title{Create or test for objects of type netSharpLayer.}
\usage{
is.netSharpLayer(x)

as.netSharpLayer(x, name, shape)
}
\arguments{
\item{x}{Object}

\item{name}{Name of the layer}

\item{shape}{Numeric vector describing the number of inputs in each dimension, e.g. \code{c(3, 15, 15)}}
}
\description{
Create or test for objects of type netSharpLayer.
}

