% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructNetDefinition.R
\name{reconstructNetDefinition}
\alias{reconstructNetDefinition}
\title{Reconstruct the NET# net definition from an neuralNet object}
\usage{
reconstructNetDefinition(modelObject, filename = NULL)
}
\arguments{
\item{modelObject}{A \link[MicrosoftML]{neuralNet} model object}

\item{filename}{Optional file name. If supplied, write model specification to file.}
}
\value{
A character vector containing the net definition
}
\description{
Reconstruct the NET# net definition from an neuralNet object
}
\examples{
library(MicrosoftML)
neuralNet(isCase ~ age + parity + education + spontaneous + induced,
                     transforms = list(isCase = case == 1),
                     data = infert,
                     numHiddenNodes = 3, 
                     verbose = 1, 
                     reportProgress = 0,
            numIterations = 20
)
model <- neuralNet(isCase ~ age + parity + education + spontaneous + induced,
                    transforms = list(isCase = case == 1),
                    data = infert,
                    numHiddenNodes = 3, 
                    verbose = 1, 
                    reportProgress = 0,
                    numIterations = 10
)
reconstructNetDefinition(model)
cat(reconstructNetDefinition(model))
tf <- tempfile(fileext = ".nn")

reconstructNetDefinition(model, filename = tf)
cat(
  readNetDefinition(tf)
)


neuralNet(isCase ~ age + parity + education + spontaneous + induced,
            transforms = list(isCase = case == 1),
            data = infert, 
            netDefinition = readNetDefinition(tf),
            numIterations = 10
            )
}
\seealso{
readNetDefinition
}

